iac/requirements.txt
aws-cdk-lib==2.181.1
python-dotenv
constructs>=10.0.0,<11.0.0
